refactor(list-pages): split flagged functions under 150 lines
Refactor the four list/content page components

- demos/DemosClient: extract useDemosData hook, demoHelpers
  (probe/params/filter-sort), DemosToolbar/DemoGridView/DemoListView
- exported-videos/ExportedVideosClient: extract useExportedVideos hook,
  videoHelpers, VideosToolbar/VideoGridView/VideoListView
- templates/TemplatesClient: extract templateData, TemplatesToolbar,
  TemplateCard/TemplateBadges/TemplateCardMeta
- preview/page: extract usePreviewPlayer hook, previewHelpers, and
  PreviewHeader/PreviewPlayer/PreviewControls/PreviewInfo/NoVideoFound
  (styled-jsx slider co-located with its input in PreviewControls)
