+++ app/(signed)/exported-videos/components/VideosToolbar.tsx
@@ -1,0 +1,92 @@
+"use client";
+import { FaTh, FaThList, FaSort, FaEye, FaListUl, FaRegClock, FaPlusSquare } from "react-icons/fa";
+import type { VideoSortOption } from "../types";
+
+interface VideosToolbarProps {
+  search: string;
+  setSearch: (value: string) => void;
+  sortDropdownOpen: boolean;
+  setSortDropdownOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  sortDropdownRef: React.RefObject<HTMLDivElement | null>;
+  sortOption: VideoSortOption;
+  setSortOption: (option: VideoSortOption) => void;
+  view: string;
+  setView: (view: string) => void;
+}
+
+export default function VideosToolbar({
+  search,
+  setSearch,
+  sortDropdownOpen,
+  setSortDropdownOpen,
+  sortDropdownRef,
+  sortOption,
+  setSortOption,
+  view,
+  setView,
+}: VideosToolbarProps) {
+  const selectSort = (option: VideoSortOption) => {
+    setSortOption(option);
+    setSortDropdownOpen(false);
+  };
+
+  const optionClass = (option: VideoSortOption) =>
+    `shared-sort-option flex items-center gap-3 text-base font-medium cursor-pointer px-2 py-2 rounded-lg hover:bg-[#F3F0FC] ${sortOption === option ? "active text-purple-700 bg-purple-50" : "text-[#A594F9]"}`;
+
+  const viewClass = (target: string) =>
+    `shared-view-btn p-3 rounded-lg border ${
+      view === target ? "active bg-[#A594F9] text-white" : "bg-white text-[#A594F9] border-gray-200"
+    }`;
+
+  return (
+    <div className="flex flex-wrap items-center justify-between mt-6 gap-4">
+      <input
+        type="text"
+        placeholder="Search your shared videos"
+        value={search}
+        onChange={(e) => setSearch(e.target.value)}
+        className="shared-search flex-1 min-w-[300px] px-4 py-3 rounded-lg bg border border-gray-200 bg-white text-gray-700 focus:outline-none focus:ring-2 focus:ring-[#A594F9]"
+      />
+
+      <div className="flex items-center gap-6 ml-auto">
+        <div className="relative">
+          <button
+            className="shared-sort-btn flex items-center gap-2 px-6 py-3 rounded-lg bg-white border border-gray-200 text-[#A594F9] font-medium hover:bg-[#ede7fa]"
+            onClick={() => setSortDropdownOpen((v) => !v)}
+          >
+            <FaSort className="text-lg" /> Sort By
+          </button>
+          {sortDropdownOpen && (
+            <div
+              ref={sortDropdownRef}
+              className="shared-sort-dropdown absolute left-0 top-full mt-2 bg-white rounded-2xl shadow-lg p-4 z-50 border border-gray-100 min-w-[180px] animate-fade-in"
+            >
+              <div className="flex flex-col gap-2">
+                <div onClick={() => selectSort("title")} className={optionClass("title")}>
+                  <FaListUl className="text-lg" /> Title
+                </div>
+                <div onClick={() => selectSort("updatedAt")} className={optionClass("updatedAt")}>
+                  <FaRegClock className="text-lg" /> Last Updated
+                </div>
+                <div onClick={() => selectSort("createdAt")} className={optionClass("createdAt")}>
+                  <FaPlusSquare className="text-lg" /> Created date
+                </div>
+                <div onClick={() => selectSort("views")} className={optionClass("views")}>
+                  <FaEye className="text-lg" /> Views
+                </div>
+              </div>
+            </div>
+          )}
+        </div>
+        <div className="flex gap-2">
+          <button className={viewClass("grid")} onClick={() => setView("grid")}>
+            <FaTh className="text-xl" />
+          </button>
+          <button className={viewClass("list")} onClick={() => setView("list")}>
+            <FaThList className="text-xl" />
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+}
