+++ app/(signed)/templates/components/TemplateCard.tsx
@@ -1,0 +1,50 @@
+import type { Template } from "../templateData";
+import TemplateBadges from "./TemplateBadges";
+import TemplateCardMeta from "./TemplateCardMeta";
+
+interface TemplateCardProps {
+  tpl: Template;
+}
+
+export default function TemplateCard({ tpl }: TemplateCardProps) {
+  return (
+    <div className="bg-white rounded-2xl p-3 sm:p-4 md:p-6 flex flex-col shadow-sm border border-[#F3F0FC]">
+      <TemplateBadges tpl={tpl} />
+      <div className="font-semibold text-base sm:text-lg text-[#1A0033] mb-1">{tpl.title}</div>
+      <div className="text-[#8B8B8B] text-xs sm:text-sm mb-3 sm:mb-4">{tpl.description}</div>
+      <div className="flex-1 flex items-center justify-center bg-[#F8F6FF] rounded-xl mb-4 sm:mb-6 min-h-[100px] sm:min-h-[120px]">
+        <svg
+          width="71"
+          height="71"
+          viewBox="0 0 71 71"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+        >
+          <mask id="path-1-inside-1_253_707" fill="white">
+            <path
+              fillRule="evenodd"
+              clipRule="evenodd"
+              d="M37.2541 67.7256L37.2215 67.7315L37.0115 67.8351L36.9523 67.8469L36.9109 67.8351L36.7009 67.7315C36.6693 67.7217 36.6456 67.7266 36.6299 67.7463L36.618 67.7759L36.5677 69.0421L36.5825 69.1012L36.6121 69.1397L36.9198 69.3586L36.9641 69.3704L36.9996 69.3586L37.3073 69.1397L37.3428 69.0924L37.3546 69.0421L37.3044 67.7789C37.2965 67.7473 37.2797 67.7296 37.2541 67.7256ZM38.038 67.3913L37.9996 67.3972L37.4523 67.6724L37.4227 67.7019L37.4138 67.7345L37.4671 69.0066L37.4819 69.0421L37.5055 69.0628L38.1001 69.3379C38.1376 69.3478 38.1662 69.3399 38.1859 69.3142L38.1978 69.2728L38.0972 67.4564C38.0873 67.4209 38.0676 67.3992 38.038 67.3913ZM35.9228 67.3972C35.9098 67.3893 35.8942 67.3868 35.8793 67.3901C35.8644 67.3934 35.8514 67.4023 35.8429 67.415L35.8252 67.4564L35.7246 69.2728C35.7266 69.3083 35.7433 69.332 35.7749 69.3438L35.8193 69.3379L36.4139 69.0628L36.4435 69.0391L36.4553 69.0066L36.5056 67.7345L36.4967 67.699L36.4671 67.6694L35.9228 67.3972Z"
+            />
+          </mask>
+          <path
+            d="M37.2541 67.7256L37.8623 63.7721L37.1988 63.67L36.5383 63.7902L37.2541 67.7256ZM37.2215 67.7315L36.5057 63.7961L35.955 63.8963L35.4529 64.1438L37.2215 67.7315ZM37.0115 67.8351L37.7959 71.7574L38.3099 71.6546L38.7801 71.4228L37.0115 67.8351ZM36.9523 67.8469L35.8535 71.693L36.7859 71.9594L37.7367 71.7692L36.9523 67.8469ZM36.9109 67.8351L35.1423 71.4228L35.4655 71.5822L35.8121 71.6812L36.9109 67.8351ZM36.7009 67.7315L38.4695 64.1438L38.1907 64.0063L37.894 63.9136L36.7009 67.7315ZM36.6299 67.7463L33.5064 65.2475L33.1361 65.7105L32.9159 66.2609L36.6299 67.7463ZM36.618 67.7759L32.9041 66.2905L32.6485 66.9295L32.6212 67.6171L36.618 67.7759ZM36.5677 69.0421L32.5709 68.8833L32.5481 69.4561L32.6872 70.0122L36.5677 69.0421ZM36.5825 69.1012L32.702 70.0714L32.904 70.8797L33.412 71.5401L36.5825 69.1012ZM36.6121 69.1397L33.4416 71.5785L33.806 72.0523L34.2931 72.3989L36.6121 69.1397ZM36.9198 69.3586L34.6008 72.6178L35.1902 73.0372L35.8892 73.2236L36.9198 69.3586ZM36.9641 69.3704L35.9336 73.2354L37.0918 73.5442L38.229 73.1652L36.9641 69.3704ZM36.9996 69.3586L38.2645 73.1534L38.8316 72.9644L39.3187 72.6178L36.9996 69.3586ZM37.3073 69.1397L39.6263 72.3989L40.1337 72.0378L40.5074 71.5396L37.3073 69.1397ZM37.3428 69.0924L40.5429 71.4923L41.0449 70.8229L41.2365 70.0084L37.3428 69.0924ZM37.3546 69.0421L41.2483 69.9582L41.3732 69.4276L41.3515 68.8829L37.3546 69.0421ZM37.3044 67.7789L41.3012 67.6197L41.2848 67.2082L41.1849 66.8087L37.3044 67.7789ZM38.038 67.3913L39.0687 63.5264L38.2587 63.3104L37.4301 63.4378L38.038 67.3913ZM37.9996 67.3972L37.3916 63.4437L36.7673 63.5397L36.203 63.8234L37.9996 67.3972ZM37.4523 67.6724L35.6557 64.0985L35.0797 64.3881L34.6238 64.8439L37.4523 67.6724ZM37.4227 67.7019L34.5943 64.8735L33.8431 65.6247L33.5636 66.6497L37.4227 67.7019ZM37.4138 67.7345L33.5547 66.6822L33.3913 67.2813L33.4173 67.9018L37.4138 67.7345ZM37.4671 69.0066L33.4706 69.1739L33.5004 69.8866L33.7748 70.5452L37.4671 69.0066ZM37.4819 69.0421L33.7896 70.5807L34.1476 71.4397L34.8479 72.0524L37.4819 69.0421ZM37.5055 69.0628L34.8716 72.0732L35.3042 72.4517L35.8258 72.693L37.5055 69.0628ZM38.1001 69.3379L36.4205 72.9681L36.7408 73.1164L37.0822 73.2062L38.1001 69.3379ZM38.1859 69.3142L41.3564 71.7531L41.8222 71.1476L42.0321 70.4131L38.1859 69.3142ZM38.1978 69.2728L42.0439 70.3716L42.2289 69.7241L42.1917 69.0517L38.1978 69.2728ZM38.0972 67.4564L42.0911 67.2352L42.0671 66.803L41.9513 66.3858L38.0972 67.4564ZM35.9228 67.3972L33.8497 70.8181L33.9886 70.9023L34.1339 70.9749L35.9228 67.3972ZM35.8429 67.415L32.5161 65.1942L32.3115 65.5006L32.1664 65.8392L35.8429 67.415ZM35.8252 67.4564L32.1487 65.8806L31.8704 66.5299L31.8313 67.2352L35.8252 67.4564ZM35.7246 69.2728L31.7307 69.0517L31.7185 69.2732L31.7308 69.4947L35.7246 69.2728ZM35.7749 69.3438L34.3704 73.0891L35.3096 73.4413L36.3038 73.3087L35.7749 69.3438ZM35.8193 69.3379L36.3482 73.3028L36.9489 73.2227L37.4989 72.9682L35.8193 69.3379ZM36.4139 69.0628L38.0936 72.693L38.5338 72.4893L38.9125 72.1864L36.4139 69.0628ZM36.4435 69.0391L38.9421 72.1627L39.819 71.4613L40.2027 70.406L36.4435 69.0391ZM36.4553 69.0066L40.2145 70.3735L40.4276 69.7876L40.4522 69.1646L36.4553 69.0066ZM36.5056 67.7345L40.5025 67.8925L40.5251 67.32L40.3861 66.7642L36.5056 67.7345ZM36.4967 67.699L40.3773 66.7287L40.1085 65.6539L39.3252 64.8706L36.4967 67.699ZM36.4671 67.6694L38.256 64.0917L37.7117 63.8195L35.9228 67.3972L34.1339 70.9749L34.6783 71.2471L36.4671 67.6694Z"
+            fill="#6356D7"
+            fillOpacity="0.53"
+            mask="url(#path-1-inside-1_253_707)"
+          />
+          <path
+            d="M20.9512 12.25C21.3159 12.2285 21.681 12.2923 22.0166 12.4365L22.0195 12.4385C25.1036 13.7569 32.0712 16.9137 40.9229 22.0225C49.7747 27.1326 55.9941 31.5904 58.6768 33.5986V33.5996C59.9431 34.5494 59.9463 36.4183 58.6787 37.3711C56.0283 39.3632 49.8856 43.7627 40.9229 48.9404C31.9513 54.1181 25.0674 57.2369 22.0166 58.5361L22.0146 58.5371C20.5629 59.1572 18.9464 58.2238 18.7568 56.6543C18.3561 53.3379 17.5996 45.7483 17.5996 35.4844C17.5996 25.226 18.3534 17.6397 18.7568 14.3223V14.3174C18.8 13.9546 18.9278 13.6068 19.1289 13.3018C19.33 12.9969 19.5996 12.7431 19.916 12.5605C20.2323 12.3781 20.5867 12.2715 20.9512 12.25Z"
+            stroke="#6356D7"
+            strokeOpacity="0.53"
+            strokeWidth="4"
+          />
+        </svg>
+      </div>
+      <TemplateCardMeta tpl={tpl} />
+      <button className="bg-[#A594F9] text-white rounded-lg px-4 sm:px-6 py-2 sm:py-3 w-full text-xs sm:text-base lg:text-lg font-medium flex items-center justify-center gap-2 mt-auto hover:bg-[#9280FF] transition">
+        Use Template
+      </button>
+    </div>
+  );
+}
